move CssVarsProvider to Joy demos
