+++ docs/src/modules/components/MarkdownDocs.js
@@ -4,13 +4,12 @@ import path from 'path';
 import { useRouter } from 'next/router';
 import { useTheme } from '@mui/system';
 import { exactProp } from '@mui/utils';
-import { CssVarsProvider, useColorScheme } from '@mui/joy/styles';
+import { CssVarsProvider } from '@mui/joy/styles';
 import Demo from 'docs/src/modules/components/Demo';
 import MarkdownElement from 'docs/src/modules/components/MarkdownElement';
 import { pathnameToLanguage } from 'docs/src/modules/utils/helpers';
 import AppLayoutDocs from 'docs/src/modules/components/AppLayoutDocs';
 import { useTranslate, useUserLanguage } from 'docs/src/modules/utils/i18n';
-import BrandingProvider from 'docs/src/BrandingProvider';
 import Ad from 'docs/src/modules/components/Ad';
 import AdGuest from 'docs/src/modules/components/AdGuest';
 
@@ -19,18 +18,6 @@ function noComponent(moduleID) {
     throw new Error(`No demo component provided for '${moduleID}'`);
   };
 }
-
-function JoyModeObserver({ mode }) {
-  const { setMode } = useColorScheme();
-  React.useEffect(() => {
-    setMode(mode);
-  }, [mode, setMode]);
-  return null;
-}
-
-JoyModeObserver.propTypes = {
-  mode: PropTypes.oneOf(['light', 'dark']),
-};
 
 export default function MarkdownDocs(props) {
   const theme = useTheme();
@@ -49,11 +36,10 @@ export default function MarkdownDocs(props) {
   const t = useTranslate();
 
   const localizedDoc = docs[userLanguage] || docs.en;
-  const { description, location, rendered, title, toc } = localizedDoc;
+  const { description, location, rendered, title, toc, headers } = localizedDoc;
 
   const isJoy = canonicalAs.startsWith('/joy-ui/');
-  const Provider = isJoy ? CssVarsProvider : React.Fragment;
-  const Wrapper = isJoy ? BrandingProvider : React.Fragment;
+  // const isCssVarsProvider = headers.provider === 'CssVarsProvider';
 
   return (
     <AppLayoutDocs
@@ -64,92 +50,81 @@ export default function MarkdownDocs(props) {
       title={title}
       toc={toc}
     >
-      <Provider>
-        {disableAd ? null : (
-          <Wrapper>
-            <AdGuest>
-              <Ad />
-            </AdGuest>
-          </Wrapper>
-        )}
-        {isJoy && <JoyModeObserver mode={theme.palette.mode} />}
-        {rendered.map((renderedMarkdownOrDemo, index) => {
-          if (typeof renderedMarkdownOrDemo === 'string') {
-            return (
-              <Wrapper key={index} {...(isJoy && { mode: theme.palette.mode })}>
-                <MarkdownElement renderedMarkdown={renderedMarkdownOrDemo} />
-              </Wrapper>
-            );
+      {isJoy && <CssVarsProvider />}
+      {disableAd ? null : (
+        <AdGuest>
+          <Ad />
+        </AdGuest>
+      )}
+      {rendered.map((renderedMarkdownOrDemo, index) => {
+        if (typeof renderedMarkdownOrDemo === 'string') {
+          return <MarkdownElement key={index} renderedMarkdown={renderedMarkdownOrDemo} />;
+        }
+
+        if (renderedMarkdownOrDemo.component) {
+          const name = renderedMarkdownOrDemo.component;
+          const Component = srcComponents?.[name];
+
+          if (Component === undefined) {
+            throw new Error(`No component found at the path ${path.join('docs/src', name)}`);
           }
 
-          if (renderedMarkdownOrDemo.component) {
-            const name = renderedMarkdownOrDemo.component;
-            const Component = srcComponents?.[name];
+          return <Component key={index} {...renderedMarkdownOrDemo} markdown={localizedDoc} />;
+        }
 
-            if (Component === undefined) {
-              throw new Error(`No component found at the path ${path.join('docs/src', name)}`);
-            }
+        const name = renderedMarkdownOrDemo.demo;
+        const demo = demos?.[name];
+        if (demo === undefined) {
+          const errorMessage = [
+            `Missing demo: ${name}. You can use one of the following:`,
+            Object.keys(demos),
+          ].join('\n');
 
-            return (
-              <Wrapper key={index} {...(isJoy && { mode: theme.palette.mode })}>
-                <Component {...renderedMarkdownOrDemo} markdown={localizedDoc} />
-              </Wrapper>
-            );
+          if (userLanguage === 'en') {
+            throw new Error(errorMessage);
           }
 
-          const name = renderedMarkdownOrDemo.demo;
-          const demo = demos?.[name];
-          if (demo === undefined) {
-            const errorMessage = [
-              `Missing demo: ${name}. You can use one of the following:`,
-              Object.keys(demos),
-            ].join('\n');
-
-            if (userLanguage === 'en') {
-              throw new Error(errorMessage);
-            }
-
-            if (process.env.NODE_ENV !== 'production') {
-              console.error(errorMessage);
-            }
-
-            const warnIcon = (
-              <span role="img" aria-label={t('emojiWarning')}>
-                ⚠️
-              </span>
-            );
-            return (
-              <div key={index}>
-                {/* eslint-disable-next-line material-ui/no-hardcoded-labels */}
-                {warnIcon} Missing demo `{name}` {warnIcon}
-              </div>
-            );
+          if (process.env.NODE_ENV !== 'production') {
+            console.error(errorMessage);
           }
 
-          const splitLocationBySlash = location.split('/');
-          splitLocationBySlash.pop();
-          const fileNameWithLocation = `${splitLocationBySlash.join('/')}/${name}`;
-
-          return (
-            <Demo
-              key={index}
-              mode={theme.palette.mode}
-              demo={{
-                raw: demo.raw,
-                js: demoComponents[demo.module] ?? noComponent(demo.module),
-                scope: demos.scope,
-                jsxPreview: demo.jsxPreview,
-                rawTS: demo.rawTS,
-                tsx: demoComponents[demo.moduleTS] ?? null,
-                gaLabel: fileNameWithLocation.replace(/^\/docs\/data\//, ''),
-              }}
-              disableAd={disableAd}
-              demoOptions={renderedMarkdownOrDemo}
-              githubLocation={`${process.env.SOURCE_CODE_REPO}/blob/v${process.env.LIB_VERSION}${fileNameWithLocation}`}
-            />
+          const warnIcon = (
+            <span role="img" aria-label={t('emojiWarning')}>
+              ⚠️
+            </span>
           );
-        })}
-      </Provider>
+          return (
+            <div key={index}>
+              {/* eslint-disable-next-line material-ui/no-hardcoded-labels */}
+              {warnIcon} Missing demo `{name}` {warnIcon}
+            </div>
+          );
+        }
+
+        const splitLocationBySlash = location.split('/');
+        splitLocationBySlash.pop();
+        const fileNameWithLocation = `${splitLocationBySlash.join('/')}/${name}`;
+
+        return (
+          <Demo
+            key={index}
+            headers={headers}
+            mode={theme.palette.mode}
+            demo={{
+              raw: demo.raw,
+              js: demoComponents[demo.module] ?? noComponent(demo.module),
+              scope: demos.scope,
+              jsxPreview: demo.jsxPreview,
+              rawTS: demo.rawTS,
+              tsx: demoComponents[demo.moduleTS] ?? null,
+              gaLabel: fileNameWithLocation.replace(/^\/docs\/data\//, ''),
+            }}
+            disableAd={disableAd}
+            demoOptions={renderedMarkdownOrDemo}
+            githubLocation={`${process.env.SOURCE_CODE_REPO}/blob/v${process.env.LIB_VERSION}${fileNameWithLocation}`}
+          />
+        );
+      })}
     </AppLayoutDocs>
   );
 }
